examples/jwt/requirements.txt
..[swagger-ui]
python-jose[cryptography]
Flask>=0.10.1
